Move these out of guice package.
