+++ gwtp-crawler-service/src/main/java/com/gwtplatform/crawlerservice/server/ServiceKey.java
@@ -20,7 +20,6 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
 import com.google.inject.BindingAnnotation;
-import com.gwtplatform.crawlerservice.server.guice.CrawlServiceServlet;
 
 import static java.lang.annotation.ElementType.FIELD;
 import static java.lang.annotation.ElementType.METHOD;
