+++ gwtp-crawler-service/src/main/java/com/gwtplatform/crawlerservice/server/guice/CrawlServiceModule.java
@@ -22,6 +22,7 @@ import com.gargoylesoftware.htmlunit.BrowserVersion;
 import com.gargoylesoftware.htmlunit.WebClient;
 import com.google.inject.Provides;
 import com.google.inject.servlet.ServletModule;
+import com.gwtplatform.crawlerservice.server.CrawlServiceServlet;
 
 public class CrawlServiceModule extends ServletModule {
